New hashdir layout and simplified updating logic.
The old code did a byte-for-byte file comparison of every file it
wanted to link into the link dirs, which was very wasteful.  The new
code incorporates the file length into the filename and relies on the
good-enough uniqueness of the md5 sum + file-length.

The old directory structure started with an 8-char prefix of the hash,
which is pretty much unique for a normal set of files, making the hash
buckets essentially useless.  The new code uses 3-char hash buckets,
3-dirs deep, and puts any overflow names (if any files reach max links)
into the final hash dir (not a dir dedicated to each hash).

For instance, given file of length 54321 and its 32-char hex hash:

Was: 12345678/901234567890123456789012/0
Now: 123/456/789/01234567890123456789012.54321.0

I also fixed the sanitizing of link-by-hash dir into a daemon module.