Improve SQLite support by unifying DB updating.
Since SQLite's locking would often fail if both the receiver and the
generator (let alone the sender too) were trying to write out new
checksums, rsync now forwards updated checksum info to a single process
(per DB).  For a local transfer (with the same DB on both sides), that
means everyone forwards to the sender proc.  For a remote transfer (or
a different DB for sending/receiving), the receiver proc forwards to
the generator proc.